+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -  **Bolt** Tests are built in batch and binaries are ran in parallel in order to speed up test times
 -  **Bolt** Run tests inside of Kubernetes pod inside cluster, removing the need for port forwarding for tests
 -  **Bolt** Remove `disable_cargo_workspace` flag since it is seldom used
+-  **Bolt** Remove  `skip_dependencies`, `force_build`, and `skip_generate` on `bolt up` and `bolt test` commands that are no longer relevant
 
 ## [23.1.0-rc4] - 2023-09-02